backend/requirements.txt
# Web framework
flask==3.1.0
flask-cors==5.0.0
Werkzeug==3.1.3

# Production server
gunicorn==23.0.0

# Configuration
python-dotenv==1.0.1

# Data processing
numpy>=1.26.0,<2.0.0
pandas>=2.2.0,<3.0.0

# Machine Learning - PyTorch
torch>=2.5.0,<3.0.0
torchaudio>=2.5.0,<3.0.0

# Machine Learning - Transformers
transformers>=4.40.0,<5.0.0
huggingface_hub>=0.26.0,<1.0.0

# Audio processing
soundfile>=0.12.1,<1.0.0

# Deepgram SDK for transcription and TTS
deepgram-sdk>=3.0.0,<4.0.0

# AI/LLM
google-generativeai>=0.8.3,<1.0.0

# Visualization (if needed for backend)
plotly>=5.24.0,<6.0.0
